Move showProjectDetails computed property to deliveries/show/index
This property is not relevant to all descendants of can-resend-controller, so I'm moving it to the appropriate location
